Update README.md
A research project focused on enhancing the efficiency of scoring patient notes written by medical students during their clinical skills assessments. Due to the extensive length of these notes and the high volume that needs to be evaluated, medical professionals often face significant time constraints and challenges in interpretation. To address this issue, the researchers developed a Named Entity Recognition (NER) technique utilizing advanced Natural Language Processing (NLP) and Transformer models. The purpose of this approach is to distill patient notes into their most crucial sentences and keywords, thereby streamlining the review process. The study experimented with two different methods, ultimately finding that the Deberta Model was superior, achieving a precision of 0.84, a recall of 0.89, and an F1-score of 0.866. This indicates that the selected method could significantly enhance the efficiency of medical documentation review.